backend/requirements.txt
fastapi==0.111.1
uvicorn==0.23.2
python-dotenv==1.0.1
sentence-transformers==2.2.2
qdrant-client==1.2.1
google-generativeai==0.3.3
